+++ 05-JWT-Basics/starter/controllers/main.js
@@ -1,5 +1,4 @@
-const jwt = require("jsonwebtoken");
-const CustomAPIError = require("../errors/custom-error");
+const jwt = require('jsonwebtoken')
 const login = async (req, res) => {
   const { username, password } = req.body;
   //Mongoose validation
@@ -24,28 +23,12 @@ const login = async (req, res) => {
 };
 
 const dashboard = async (req, res) => {
-  const {authorization} = req.headers;
-  console.log("Authorization", authorization);
-  console.log(req.headers);
-  if (!authorization || !authorization.startsWith("Bearer ")) {
-    throw new CustomAPIError("No token provided", 401);
-  }
-  
-  const token = authorization.split(" ")[1];
-  console.log(token)
-  try {
-    const decoded = jwt.verify(token,process.env.JWT_SECRET)
-    console.log(decoded)  
-    const luckyNumber = Math.floor(Math.random() * 100);
+
+  const luckyNumber = Math.floor(Math.random() * 100);
   res.status(200).json({
-    msg: `Hello ${decoded.username}`,
+    msg: `Hello ${req.user.username}`,
     secret: `Here is your authorized data, your lucky number is ${luckyNumber}`,
   });
-  } catch (error) {
-    
-    throw new CustomAPIError("Not authorizad to access this route", 401);
-    
-  }
 
   
 };
